+++ NEWS.md
@@ -1,4 +1,4 @@
-# zip (development version)
+# zip 2.3.0
 
 * zip now handles large zip files on Windows (#65, #75, #79, @weshinsley).